+++ app/src/main/java/com/yuno/payments/example/ui/enrollment/EnrollPaymentActivity.kt
@@ -1,12 +1,14 @@
 package com.yuno.payments.example.ui.enrollment
 
 import android.os.Bundle
+import android.util.Log
 import android.widget.Button
 import android.widget.EditText
 import androidx.appcompat.app.AppCompatActivity
 import com.yuno.payments.example.BuildConfig
 import com.yuno.payments.example.R
 import com.yuno.payments.example.ui.views.CustomEditText
+import com.yuno.payments.features.enrollment.initEnrollment
 import com.yuno.payments.features.enrollment.startEnrollment
 
 class EnrollPaymentActivity : AppCompatActivity() {
@@ -19,6 +21,7 @@ class EnrollPaymentActivity : AppCompatActivity() {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_enroll_payment)
         initViews()
+        initEnrollment(::onEnrollmentStateChange)
     }
 
     private fun initViews() {
@@ -38,6 +41,12 @@ class EnrollPaymentActivity : AppCompatActivity() {
                 customerSession = editTextCustomerSession.text.toString(),
                 countryCode = editTextCountryCode.text.toString()
             )
+        }
+    }
+
+    private fun onEnrollmentStateChange(enrollmentState: String?) {
+        enrollmentState?.let {
+            Log.e("Enrollment State", it)
         }
     }
 }